pyproject.toml
[project]
name = "fitbod-report"
version = "0.2.0"
description = "A Fitbod workout report generator application"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.42.0",
    "markdown-pdf>=1.13.0",
    "watchdog>=6.0.0",
    "tzlocal>=5.2",
    "pyyaml>=6.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.9.0"
]

[tool.pytest.ini_options]
pythonpath = ["tests"]

[tool.ruff]
line-length = 127
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "S"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101"]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
